fix(release upload): fix asset URL format for uploads in GitLab 17.2
GitLab enforces the use of the new URL format for uploads created
starting 17.2. So far the cli used the projectFile.URL property and
manually created the full URL using the project path. This schema
stopped working in GitLab 17.2, which enforces the use of the new upload
schema using the actual project-id number.

This commit fixes the construction of the URL by making use of the
FullPath property of projectFile, which provides backwards
compatibility for older GitLab instances and fully utilizes the actual
full path that the GitLab API upload call returns.

Caused-by: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/156939
Related-to: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150939
Signed-off-by: Levente Polyak <levente@leventepolyak.net>