fix(issueable list): remove support multiple not- filter flags
This change set fixes the issueable list command line interface by
removing support for providing multiple `--not-assignee` and
`not-author` flags. Those two flags have never functioned the way it was
intended, instead they silently failed by only considering the first
passed value to the API. All other values have been ignored.

This change set therefore changes so that the `--not-assignee` and
`--not-author` flags can only submitted once.

This is a breaking change to the CLI justified by the feature never
working as intended without providing any information to the user.

An alternative fix would submit the API request multiple times and merge
the results.