refactor: always use factory to create client from commands
This change set makes use of the new `ApiClient` function in the factory
instead of directly creating a new client. This has the benefit that
testing has a smaller interface of just depending on the factory and
nothing else. Without this and without the global api client this would
be hard.