feat(job artifact): Add --list-paths to print paths
I've been confused by this a few times when the artifacts are under .gitignored
paths. The command currently provides no output on success, so this flag adds
the output of the path of each file as it's copied from the Zip.

As `ci artifact` is deprecated, I've omitted the flag there and pass the default
value `false`.

Issue: https://gitlab.com/gitlab-org/cli/-/issues/7914