fix(repo list): fix project group filtering reliability
The issue was that the previous implementation relied on the group search which could match multiple groups. This was accounted for by manually filtering on FullPath; however what was not covered was pagination. If the group with the correct fullpath was not on the first page of results, it would result in "No group matching path .." error.
The fix is to use the group detail endpoint instead.
Also fixes the test data as the full path of groups does not include a leading /

Fixes #7615