feat(duo claude): add option to configure GitLab MCP for claude code
This change set adds the `--with-gitlab-mcp` option to the `glab duo
claude` command to configure it with the official GitLab MCP server.

It currently uses the HTTP transport method, which is not yet fully
supported in Claude Code with GitLab, because of missing support in
Claude Code for OAuth2 scopes during the dynamic client registration.
GitLab requires the `mcp` scope here.

An upstream issue exists for it in:
https://github.com/anthropics/claude-code/issues/4540

To work around this, we use `npx mcp-remote` if `npx` is available.