feat: sign all macOS and Windows binaries in goreleaser
This commit builds on top of
https://gitlab.com/gitlab-org/cli/-/merge_requests/2381. Ideally
we would use goreleaser to build **all** binaries and extract
what we need, but the open source version of goreleaser does not
have the ability to split the build, sign, and publishing step.

For now, we just use Docker-in-Docker to launch a code-signer image to
sign both macOS and Windows binaries. That means we have to ensure the
container has access to the required environment variables in order to
sign the binaries.

Upgrade to code-signer v1.2.1 to obtain some of the improvements for
simplifying the signing of binaries.