fix: tweak signature name to avoid bin/
When `signature` contained the artifact name, the .sha256 files
would be prefixed with `bin/glab`. When goreleaser went to upload
the signature files, GitLab would report a 400 error because
it doesn't allow names that contain `/`.

Fix this by generating the signature with the project name instead
of the artifact name.