feat(config): setting empty config value removes its node from config
We don't have a feature were an empty value has a meaning. Rather the
opposite, where for example during a `glab auth logout` we used to
simply set the token to an empty string which semantically is weird.

This change set will remove the node from the config when set to
an empty string remove.

Refs https://gitlab.com/gitlab-org/cli/-/issues/8001